requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
httpx==0.27.2
jsonpath-ng==1.6.1
google-generativeai==0.8.3
reportlab==4.2.5